requirements.txt
streamlit==1.55.0
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.2
trafilatura==2.0.0
pandas==2.3.3
openpyxl==3.1.5
Janome==0.5.0
wordcloud==1.9.6
matplotlib==3.10.8
Pillow==12.1.1
PyYAML==6.0.3
tldextract==5.3.1
